Fix several signedness warnings in gio/tests/actions.c
gio/tests/actions.c: In function ‘strv_set_equal’:
gio/tests/actions.c:177:41: error: comparison of integer expressions of different signedness: ‘guint’ {aka ‘unsigned int’} and ‘gint’ {aka ‘int’}
  177 |     res = g_strv_length ((gchar**)strv) == count;
      |                                         ^~
gio/tests/actions.c: In function ‘test_parse_detailed’:
gio/tests/actions.c:473:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’
  473 |   for (i = 0; i < G_N_ELEMENTS (testcases); i++)
      |                 ^

+++ gio/tests/actions.c
@@ -151,7 +151,7 @@ strv_strv_cmp (gchar **a, gchar **b)
 static gboolean
 strv_set_equal (gchar **strv, ...)
 {
-  gint count;
+  guint count;
   va_list list;
   const gchar *str;
   gboolean res;
@@ -468,7 +468,7 @@ test_parse_detailed (void)
     { "abc(42, 4)",       "abc",    "(42, 4)",  "expected end of input", NULL },
     { "abc(42,)",         "abc",    "(42,)",    "expected end of input", NULL }
   };
-  gint i;
+  gsize i;
 
   for (i = 0; i < G_N_ELEMENTS (testcases); i++)
     {
